04-storage-and-persistence/01-pactice/k8s/vagrant/m4/part3/docker-images/k8s-facts/code/requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
dnspython
requests